\name{.First.lib}
\alias{.First.lib}
\title{Load all packages formerly part of the 'gregmisc' bundle}
\description{
  Load all packages formerly part of the 'gregmisc' bundle
}
\usage{
.First.lib(libname, pkgname)
}
\arguments{
  \item{libname}{ignored}
  \item{pkgname}{ignored}
}
\value{
  Nothing of interest
}
\author{ Gregory R. Warnes \email{warnes@bst.rochester.edu}}
\seealso{ \code{\link[base]{.First.lib}}}
\examples{
## Should never be directly executed ##
}
\keyword{ misc }
