% $Id: undocumented.Rd,v 1.8 2004/03/26 22:27:37 warnes Exp $

\name{undocumented}
\alias{iapply}
\alias{residplot}
\title{Undocumented functions}
\description{
  These functions are undocumented.  Some are internal and not intended
  for direct use.  Others simply haven't been documented yet.  
}
\usage{
residplot(model, formula, ...) 
}
\arguments{
  \item{...}{Undocumented}
  \item{model}{Undocumented}
  \item{formula}{Undocumented}
  }
\author{Gregory R. Warnes \email{gregory\_r\_warnes@groton.pfizer.com}}
\keyword{misc}
